+++ examples/simple-client/simple_client.go
@@ -27,8 +27,10 @@ import (
 	"git.fd.io/govpp.git/adapter/socketclient"
 	"git.fd.io/govpp.git/api"
 	"git.fd.io/govpp.git/core"
+	"git.fd.io/govpp.git/examples/binapi/interface_types"
 	"git.fd.io/govpp.git/examples/binapi/interfaces"
 	"git.fd.io/govpp.git/examples/binapi/ip"
+	"git.fd.io/govpp.git/examples/binapi/ip_types"
 	"git.fd.io/govpp.git/examples/binapi/vpe"
 )
 
@@ -166,8 +168,8 @@ func addIPAddress(ch api.Channel, index interfaces.InterfaceIndex) {
 		IsAdd:     true,
 		Prefix: interfaces.AddressWithPrefix{
 			Address: interfaces.Address{
-				Af: interfaces.ADDRESS_IP4,
-				Un: interfaces.AddressUnionIP4(interfaces.IP4Address{10, 10, 0, uint8(index)}),
+				Af: ip_types.ADDRESS_IP4,
+				Un: ip_types.AddressUnionIP4(interfaces.IP4Address{10, 10, 0, uint8(index)}),
 			},
 			Len: 32,
 		},
@@ -244,7 +246,7 @@ func interfaceNotifications(ch api.Channel, index interfaces.InterfaceIndex) {
 	// generate some events in VPP
 	err = ch.SendRequest(&interfaces.SwInterfaceSetFlags{
 		SwIfIndex: index,
-		Flags:     interfaces.IF_STATUS_API_FLAG_ADMIN_UP,
+		Flags:     interface_types.IF_STATUS_API_FLAG_ADMIN_UP,
 	}).ReceiveReply(&interfaces.SwInterfaceSetFlagsReply{})
 	if err != nil {
 		logError(err, "setting interface flags")